remove public streaming stuff
